move to a config instead of env
